user request 3.3
Refact all funtion to to delete address delete with addressid and userid  for delete current address of user

+++ pkg/repository/user.repo.go
@@ -36,11 +36,11 @@ type userRepo struct {
 }
 
 // DeleteAddress implements interfaces.UserRepository
-func (c *userRepo) DeleteAddress(id int) error {
-	query := `DELETE FROM addresses WHERE id_address=$1 RETURNING id_address;`
+func (c *userRepo) DeleteAddress(id int, userid int) error {
+	query := `DELETE FROM addresses WHERE id_address=$1 AND user_id=$2 RETURNING id_address;`
 
 	var row int
-	sql := c.db.QueryRow(query, id)
+	sql := c.db.QueryRow(query, id, userid)
 
 	sql.Scan(&row)
 	if row == 0 {
